Refactor StudentResults page
Refactor the StudentResults page into smaller, reusable components for improved maintainability and readability. This includes creating separate components for the results table, chart, add form, and filters.
